pdf-it-priorities-chatbot/requirements.txt
streamlit==1.31.0
langchain==0.1.10
langchain-openai==0.0.8
pypdf2==3.0.1
faiss-cpu==1.8.0
openai==1.13.3
python-dotenv==1.0.1